webapi/requirements.txt
django==1.11.13
djangorestframework==3.8.2
django-rest-swagger==2.1.2
django-crispy-forms==1.7.2
pytz==2017.2
django-filter==1.1.0
djangorestframework-filters==0.10.2
django-cors-headers==2.1.0
drf-dynamic-fields==0.3.0
pymysql==0.8.0
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
idna==2.6
itypes==1.1.0
jinja2==2.10
openapi-codec==1.3.2
requests==2.18.4
simplejson==3.14.0
uritemplate==3.0.0
urllib3==1.22
PyJWT==1.5.3
python-dateutil==2.7.3
djangorestframework-jwt==1.11.0
#MySQL-python==1.2.5
coreapi==2.3.3
coreschema==0.0.4
<<<<<<< /home/crsadm/dev/webapi/requirements.txt
=======
django_rest_swagger==2.1.2>>>>>>> /tmp/requirements.txt~other.F4mBNr
